+++ MulliganMadness/Curses/AutoPickCurse.cs
@@ -23,6 +23,7 @@ namespace MulliganMadness.Curses
                 ExclusiveCategory
             };
             cardInfo.blacklistedCategories = new[] { ExclusiveCategory };
+            CardArtFactory.TryAssignSprite(cardInfo);
         }
 
         public override void OnAddCard(Player player, Gun gun, GunAmmo gunAmmo, CharacterData data, HealthHandler health, Gravity gravity, Block block, CharacterStatModifiers characterStats)